requirements.txt
flask
pandas
python-dotenv
markdown
pillow
groq
openpyxl
vercel-wsgi
